clearSky/requirements.txt
django
djangorestframework
scikit-learn
numpy
pandas
pickle
